repo-updater: Move PhabricatorSource to phabricator.go
