+++ cmd/repo-updater/repos/sources.go
@@ -8,11 +8,9 @@ import (
 	"sync"
 	"time"
 
-	"github.com/goware/urlx"
 	multierror "github.com/hashicorp/go-multierror"
 	"github.com/pkg/errors"
 	"github.com/sourcegraph/sourcegraph/pkg/api"
-	"github.com/sourcegraph/sourcegraph/pkg/extsvc/phabricator"
 	"github.com/sourcegraph/sourcegraph/pkg/httpcli"
 	"github.com/sourcegraph/sourcegraph/pkg/jsonc"
 	"github.com/sourcegraph/sourcegraph/schema"
@@ -180,174 +178,6 @@ func group(srcs []Source) map[string]Sources {
 	}
 
 	return groups
-}
-
-// A PhabricatorSource yields repositories from a single Phabricator connection configured
-// in Sourcegraph via the external services configuration.
-type PhabricatorSource struct {
-	svc  *ExternalService
-	conn *schema.PhabricatorConnection
-	cf   *httpcli.Factory
-
-	mu  sync.Mutex
-	cli *phabricator.Client
-}
-
-// NewPhabricatorSource returns a new PhabricatorSource from the given external service.
-func NewPhabricatorSource(svc *ExternalService, cf *httpcli.Factory) (*PhabricatorSource, error) {
-	var c schema.PhabricatorConnection
-	if err := jsonc.Unmarshal(svc.Config, &c); err != nil {
-		return nil, errors.Wrapf(err, "external service id=%d config error", svc.ID)
-	}
-	return &PhabricatorSource{svc: svc, conn: &c, cf: cf}, nil
-}
-
-// ListRepos returns all Phabricator repositories accessible to all connections configured
-// in Sourcegraph via the external services configuration.
-func (s *PhabricatorSource) ListRepos(ctx context.Context) (repos []*Repo, err error) {
-	cli, err := s.client(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	urn := s.svc.URN()
-
-	cursor := &phabricator.Cursor{Limit: 100, Order: "oldest"}
-	for {
-		var page []*phabricator.Repo
-		page, cursor, err = cli.ListRepos(ctx, phabricator.ListReposArgs{Cursor: cursor})
-		if err != nil {
-			return nil, err
-		}
-
-		for _, r := range page {
-			if r.VCS != "git" || r.Status == "inactive" {
-				continue
-			}
-
-			repo, err := phabricatorRepoToRepo(urn, s.conn, r)
-			if err != nil {
-				return nil, err
-			}
-			repos = append(repos, repo)
-		}
-
-		if cursor.After == "" {
-			break
-		}
-	}
-
-	return repos, nil
-}
-
-// ExternalServices returns a singleton slice containing the external service.
-func (s *PhabricatorSource) ExternalServices() ExternalServices {
-	return ExternalServices{s.svc}
-}
-
-func phabricatorRepoToRepo(
-	urn string,
-	conn *schema.PhabricatorConnection,
-	repo *phabricator.Repo,
-) (*Repo, error) {
-	var external []*phabricator.URI
-	builtin := make(map[string]*phabricator.URI)
-
-	for _, u := range repo.URIs {
-		if u.Disabled || u.Normalized == "" {
-			continue
-		} else if u.BuiltinIdentifier != "" {
-			builtin[u.BuiltinProtocol+"+"+u.BuiltinIdentifier] = u
-		} else {
-			external = append(external, u)
-		}
-	}
-
-	var name string
-	if len(external) > 0 {
-		name = external[0].Normalized
-	}
-
-	var cloneURL string
-	for _, alt := range [...]struct {
-		protocol, identifier string
-	}{ // Ordered by priority.
-		{"https", "shortname"},
-		{"https", "callsign"},
-		{"https", "id"},
-		{"ssh", "shortname"},
-		{"ssh", "callsign"},
-		{"ssh", "id"},
-	} {
-		if u, ok := builtin[alt.protocol+"+"+alt.identifier]; ok {
-			cloneURL = u.Effective
-			// TODO(tsenart): Authenticate the cloneURL with the user's
-			// VCS password once we have that setting in the config. The
-			// Conduit token can't be used for cloning.
-			// cloneURL = setUserinfoBestEffort(cloneURL, conn.VCSPassword, "")
-
-			if name == "" {
-				name = u.Normalized
-			}
-		}
-	}
-
-	if cloneURL == "" {
-		return nil, errors.Errorf("no clone URL available for repo with id=%v", repo.ID)
-	}
-
-	if name == "" {
-		return nil, errors.Errorf("no canonical name available for repo with id=%v", repo.ID)
-	}
-
-	serviceID, err := urlx.NormalizeString(conn.Url)
-	if err != nil {
-		// Should never happen. URL must be validated on input.
-		panic(err)
-	}
-
-	return &Repo{
-		Name: name,
-		URI:  name,
-		ExternalRepo: api.ExternalRepoSpec{
-			ID:          repo.PHID,
-			ServiceType: "phabricator",
-			ServiceID:   serviceID,
-		},
-		Enabled: true,
-		Sources: map[string]*SourceInfo{
-			urn: {
-				ID:       urn,
-				CloneURL: cloneURL,
-				// TODO(tsenart): We need a way for admins to specify which URI to
-				// use as a CloneURL. Do they want to use https + shortname, git + callsign
-				// an external URI that's mirrored or observed, etc.
-				// This must be figured out when starting to integrate the new Syncer with this
-				// source.
-			},
-		},
-		Metadata: repo,
-	}, nil
-}
-
-// client initialises the phabricator.Client if it isn't initialised yet.
-// This is done lazily instead of in NewPhabricatorSource so that we have
-// access to the context.Context passed in via ListRepos.
-func (s *PhabricatorSource) client(ctx context.Context) (*phabricator.Client, error) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	if s.cli != nil {
-		return s.cli, nil
-	}
-
-	hc, err := s.cf.Doer()
-	if err != nil {
-		return nil, err
-	}
-
-	s.cli, err = phabricator.NewClient(ctx, s.conn.Url, s.conn.Token, hc)
-	return s.cli, err
 }
 
 // A OtherSource yields repositories from a single Other connection configured
